Update horizontal card text
There are two horizontal cards in the deck. The horizontal cards are considered to have the same meaning in both the upright and reversed orientations. The "reversed" text reference the "upright" text actually providing the text from that upright card. This change adds the "upright" text of the card to the "reversed" text. 

(Changes untested)